Wire CI: host build + GD32 port syntax-checks
Three lanes in .github/workflows/papr-ci.yml run on every push / PR that
touches papr/:

  host-build       cmake build of papr_core + simulation HAL stub, then a
                   2-second smoke run of papr_firmware (124 = still alive)
  GD32E517RE       gcc -fsyntax-only on the HAL implementation and on the
  GD32E503CE        reference bootloader, against a permissive vendor-SDK
                   shim under papr/ci/gd32_shim/

The MCU lanes are a syntax-check, not a flashable build — never deploy
artifacts from them. Their purpose is to catch the most common regression
class on a multi-MCU port: refactoring the papr_hal contract and forgetting
to update one of the implementations.

ci/gd32_shim/
 - gd32_periph.h declares (does not define) every GD32 SDK function the
   HAL ports call, plus types/macros/constants. Header is portable C so
   plain gcc works; arm-none-eabi-gcc on the runner works too.
 - gd32e51x.h / gd32e50x.h are thin includes of the shared header.

Caught during bring-up of the CI and fixed in this commit:
 - hal/gd32e503ce/papr_pinmap.h: flash-map overrides now go through
   #undef so the override is explicit independent of include order, and
   PAPR_SLOT_APP_MAX / PAPR_MANIFEST_OFFSET (computed from the overridden
   macros) are also redefined.
 - boot/papr_boot.c: gated only on GD32E51X before; now compiles for
   both GD32E51X and GD32E50X (the FMC + CMSIS API is identical across
   the families). The bootloader CI lane covers both flash layouts.
 - hal/gd32e51{7,3}re/...: ota_staging_slot_index / running_slot_index
   tripped -Wconversion on the implicit unsigned->uint8_t narrowing of
   the ternary result. Cast added.

Docs
 - README: new "Continuous integration" subsection under Building.
 - ARCHITECTURE: ci/ entry in the directory map; cross-reference.

Host build still clean; all four GD32 syntax-checks (HAL + bootloader for
each MCU) clean under -Wall -Wextra -Wpedantic -Wshadow -Wconversion.

+++ papr/hal/gd32e503ce/papr_hal_gd32e503ce.c
@@ -711,11 +711,11 @@ papr_status_t papr_hal_ota_read(uint32_t offset, uint8_t *data, uint32_t len)
 
 static uint8_t ota_staging_slot_index(void)
 {
-    return (ota_staging_base() == PAPR_SLOT_A_ADDR) ? 0U : 1U;
+    return (uint8_t)((ota_staging_base() == PAPR_SLOT_A_ADDR) ? 0U : 1U);
 }
 static uint8_t ota_running_slot_index(void)
 {
-    return (ota_running_slot_base() == PAPR_SLOT_A_ADDR) ? 0U : 1U;
+    return (uint8_t)((ota_running_slot_base() == PAPR_SLOT_A_ADDR) ? 0U : 1U);
 }
 
 static void boot_state_read(papr_boot_state_record_t *r)
